+++ code/engine/viwaveformer/headers/viwaveformer.h
@@ -3,7 +3,7 @@
 
 #include "viaudiobuffer.h"
 
-typedef ViChunk<qreal> ViWaveFormChunk;
+typedef ViChunk<float> ViWaveFormChunk;
 
 class ViWaveFormer
 {
@@ -12,17 +12,17 @@ class ViWaveFormer
 		ViWaveFormer(int sampleSize = 16);
 		void setSampleSize(int size);
 		int pcmToReal(ViAudioBufferChunk *chunkIn, ViWaveFormChunk *chunkOut);
-		int pcmToReal(ViAudioBufferChunk *chunkIn, qreal *dataOut);
-		int pcmToReal(char *dataIn, qreal *dataOut, int size);
+		int pcmToReal(ViAudioBufferChunk *chunkIn, float *dataOut);
+		int pcmToReal(char *dataIn, float *dataOut, int size);
 
 	private:
-		int pcmToReal8(char* buffer, qreal *result, int size);
-		int pcmToReal16(char* buffer, qreal *result, int size);
-		int pcmToReal24(char* buffer, qreal *result, int size);
-		int pcmToReal32(char* buffer, qreal *result, int size);
+		int pcmToReal8(char* buffer, float *result, int size);
+		int pcmToReal16(char* buffer, float *result, int size);
+		int pcmToReal24(char* buffer, float *result, int size);
+		int pcmToReal32(char* buffer, float *result, int size);
 
 	private:
-		int (ViWaveFormer::*pcmToRealPointer)(char*, qreal*, int); //Function pointer
+		int (ViWaveFormer::*pcmToRealPointer)(char*, float*, int); //Function pointer
 		int mSampleSize;
 };
