+++ code/engine/viwaveformer/sources/vipcmconverter.cpp
@@ -12,47 +12,47 @@
 #define PCMS_32_MAX_VALUE 2147483647u;
 #define PCMS_32_MAX_AMPLITUDE 2147483648u;
 
-qreal ViPcmConverter::pcmToReal8(qint8 pcm)
+float ViPcmConverter::pcmToReal8(qint8 pcm)
 {
-    return qreal(pcm) / PCMS_8_MAX_AMPLITUDE;
+    return float(pcm) / PCMS_8_MAX_AMPLITUDE;
 }
 
-qint8 ViPcmConverter::realToPcm8(qreal real)
+qint8 ViPcmConverter::realToPcm8(float real)
 {
     return real * PCMS_8_MAX_VALUE;
 }
 
-qreal ViPcmConverter::pcmToReal16(qint16 pcm)
+float ViPcmConverter::pcmToReal16(qint16 pcm)
 {
-    return qreal(pcm) / PCMS_16_MAX_AMPLITUDE;
+    return float(pcm) / PCMS_16_MAX_AMPLITUDE;
 }
 
-qint16 ViPcmConverter::realToPcm16(qreal real)
+qint16 ViPcmConverter::realToPcm16(float real)
 {
     return real * PCMS_16_MAX_VALUE;
 }
 
-qreal ViPcmConverter::pcmToReal24(qint32 pcm)
+float ViPcmConverter::pcmToReal24(qint32 pcm)
 {
-    return qreal(pcm) / PCMS_24_MAX_AMPLITUDE;
+    return float(pcm) / PCMS_24_MAX_AMPLITUDE;
 }
 
-qint32 ViPcmConverter::realToPcm24(qreal real)
+qint32 ViPcmConverter::realToPcm24(float real)
 {
     return real * PCMS_24_MAX_VALUE;
 }
 
-qreal ViPcmConverter::pcmToReal32(qint32 pcm)
+float ViPcmConverter::pcmToReal32(qint32 pcm)
 {
-    return qreal(pcm) / PCMS_32_MAX_AMPLITUDE;
+    return float(pcm) / PCMS_32_MAX_AMPLITUDE;
 }
 
-qint32 ViPcmConverter::realToPcm32(qreal real)
+qint32 ViPcmConverter::realToPcm32(float real)
 {
     return real * PCMS_32_MAX_VALUE;
 }
 
-int ViPcmConverter::pcmToReal8(char* buffer, qreal *result, int size)
+int ViPcmConverter::pcmToReal8(char* buffer, float *result, int size)
 {
 	qint8 *base = reinterpret_cast<qint8*>(buffer);
 	for(int i = 0; i < size; ++i)
@@ -62,7 +62,7 @@ int ViPcmConverter::pcmToReal8(char* buffer, qreal *result, int size)
 	return size;
 }
 
-int ViPcmConverter::pcmToReal16(char* buffer, qreal *result, int size)
+int ViPcmConverter::pcmToReal16(char* buffer, float *result, int size)
 {
 	size /= 2; //char is only 8bit
 	qint16 *base = reinterpret_cast<qint16*>(buffer);
@@ -73,7 +73,7 @@ int ViPcmConverter::pcmToReal16(char* buffer, qreal *result, int size)
 	return size;
 }
 
-int ViPcmConverter::pcmToReal24(char* buffer, qreal *result, int size)
+int ViPcmConverter::pcmToReal24(char* buffer, float *result, int size)
 {
 	size /= 3; //char is only 8bit
 	qint16 *base = reinterpret_cast<qint16*>(buffer);
@@ -84,7 +84,7 @@ int ViPcmConverter::pcmToReal24(char* buffer, qreal *result, int size)
 	return size;
 }
 
-int ViPcmConverter::pcmToReal32(char* buffer, qreal *result, int size)
+int ViPcmConverter::pcmToReal32(char* buffer, float *result, int size)
 {
 	size /= 4; //char is only 8bit
 	qint32 *base = reinterpret_cast<qint32*>(buffer);
