+++ code/engine/viwaveformer/sources/viwaveformer.cpp
@@ -36,32 +36,32 @@ int ViWaveFormer::pcmToReal(ViAudioBufferChunk *chunkIn, ViWaveFormChunk *chunkO
 	return pcmToReal(chunkIn->data(), chunkOut->data(), chunkIn->size());
 }
 
-int ViWaveFormer::pcmToReal(ViAudioBufferChunk *chunkIn, qreal *dataOut)
+int ViWaveFormer::pcmToReal(ViAudioBufferChunk *chunkIn, float *dataOut)
 {
 	return pcmToReal(chunkIn->data(), dataOut, chunkIn->size());
 }
 
-int ViWaveFormer::pcmToReal(char *dataIn, qreal *dataOut, int size)
+int ViWaveFormer::pcmToReal(char *dataIn, float *dataOut, int size)
 {
 	return (this->*pcmToRealPointer)(dataIn, dataOut, size);
 }
 
-int ViWaveFormer::pcmToReal8(char* buffer, qreal *result, int size)
+int ViWaveFormer::pcmToReal8(char* buffer, float *result, int size)
 {
 	return ViPcmConverter::pcmToReal8(buffer, result, size);
 }
 
-int ViWaveFormer::pcmToReal16(char* buffer, qreal *result, int size)
+int ViWaveFormer::pcmToReal16(char* buffer, float *result, int size)
 {
 	return ViPcmConverter::pcmToReal16(buffer, result, size);
 }
 
-int ViWaveFormer::pcmToReal24(char* buffer, qreal *result, int size)
+int ViWaveFormer::pcmToReal24(char* buffer, float *result, int size)
 {
 	return ViPcmConverter::pcmToReal24(buffer, result, size);
 }
 
-int ViWaveFormer::pcmToReal32(char* buffer, qreal *result, int size)
+int ViWaveFormer::pcmToReal32(char* buffer, float *result, int size)
 {
 	return ViPcmConverter::pcmToReal32(buffer, result, size);
 }
